+++ src/model/BaseDeal.js
@@ -434,7 +434,7 @@ BaseDeal.prototype['visible_to'] = undefined;
 BaseDeal.prototype['close_time'] = undefined;
 
 /**
- * The ID of pipeline associated with the deal
+ * The ID of the pipeline associated with the deal
  * @member {Number} pipeline_id
  */
 BaseDeal.prototype['pipeline_id'] = undefined;