+++ src/model/DealNonStrict.js
@@ -495,7 +495,7 @@ DealNonStrict.prototype['visible_to'] = undefined;
 DealNonStrict.prototype['close_time'] = undefined;
 
 /**
- * The ID of pipeline associated with the deal
+ * The ID of the pipeline associated with the deal
  * @member {Number} pipeline_id
  */
 DealNonStrict.prototype['pipeline_id'] = undefined;
@@ -812,7 +812,7 @@ BaseDeal.prototype['visible_to'] = undefined;
  */
 BaseDeal.prototype['close_time'] = undefined;
 /**
- * The ID of pipeline associated with the deal
+ * The ID of the pipeline associated with the deal
  * @member {Number} pipeline_id
  */
 BaseDeal.prototype['pipeline_id'] = undefined;