+++ src/model/NewDealProduct.js
@@ -13,7 +13,7 @@
 
 import ApiClient from '../ApiClient';
 import BasicDealProduct from './BasicDealProduct';
-import NumberBoolean from './NumberBoolean';
+import DealProductUnitDuration from './DealProductUnitDuration';
 
 /**
  * The NewDealProduct model module.
@@ -82,6 +82,11 @@ class NewDealProduct {
 
                 delete data['duration'];
             }
+            if (data.hasOwnProperty('duration_unit')) {
+                obj['duration_unit'] = ApiClient.convertToType(data['duration_unit'], DealProductUnitDuration);
+
+                delete data['duration_unit'];
+            }
             if (data.hasOwnProperty('product_variation_id')) {
                 obj['product_variation_id'] = ApiClient.convertToType(data['product_variation_id'], 'Number');
 
@@ -98,7 +103,7 @@ class NewDealProduct {
                 delete data['tax'];
             }
             if (data.hasOwnProperty('enabled_flag')) {
-                obj['enabled_flag'] = ApiClient.convertToType(data['enabled_flag'], NumberBoolean);
+                obj['enabled_flag'] = ApiClient.convertToType(data['enabled_flag'], 'Boolean');
 
                 delete data['enabled_flag'];
             }
@@ -140,11 +145,17 @@ NewDealProduct.prototype['quantity'] = undefined;
 NewDealProduct.prototype['discount_percentage'] = 0;
 
 /**
- * The duration of the product (when product durations are not enabled for the company or if omitted, defaults to 1)
+ * The duration of the product. If omitted, will be set to 1.
  * @member {Number} duration
  * @default 1
  */
 NewDealProduct.prototype['duration'] = 1;
+
+/**
+ * The unit duration of the product
+ * @member {module:model/DealProductUnitDuration} duration_unit
+ */
+NewDealProduct.prototype['duration_unit'] = undefined;
 
 /**
  * The ID of the product variation to use. When omitted, no variation will be used.
@@ -153,7 +164,7 @@ NewDealProduct.prototype['duration'] = 1;
 NewDealProduct.prototype['product_variation_id'] = undefined;
 
 /**
- * Any textual comment associated with this product-deal attachment. Visible and editable in the application UI.
+ * A textual comment associated with this product-deal attachment
  * @member {String} comments
  */
 NewDealProduct.prototype['comments'] = undefined;
@@ -166,10 +177,11 @@ NewDealProduct.prototype['comments'] = undefined;
 NewDealProduct.prototype['tax'] = 0;
 
 /**
- * Whether the product is enabled on the deal or not. This makes it possible to add products to a deal with a specific price and discount criteria - but keep them disabled, which refrains them from being included in the deal price calculation. When omitted, the product will be marked as enabled by default.
- * @member {module:model/NumberBoolean} enabled_flag
+ * Whether the product is enabled for a deal or not. This makes it possible to add products to a deal with a specific price and discount criteria, but keep them disabled, which refrains them from being included in the deal value calculation. When omitted, the product will be marked as enabled by default.
+ * @member {Boolean} enabled_flag
+ * @default true
  */
-NewDealProduct.prototype['enabled_flag'] = undefined;
+NewDealProduct.prototype['enabled_flag'] = true;
 
 
 // Implement BasicDealProduct interface:
@@ -195,18 +207,23 @@ BasicDealProduct.prototype['quantity'] = undefined;
  */
 BasicDealProduct.prototype['discount_percentage'] = 0;
 /**
- * The duration of the product (when product durations are not enabled for the company or if omitted, defaults to 1)
+ * The duration of the product. If omitted, will be set to 1.
  * @member {Number} duration
  * @default 1
  */
 BasicDealProduct.prototype['duration'] = 1;
+/**
+ * The unit duration of the product
+ * @member {module:model/DealProductUnitDuration} duration_unit
+ */
+BasicDealProduct.prototype['duration_unit'] = undefined;
 /**
  * The ID of the product variation to use. When omitted, no variation will be used.
  * @member {Number} product_variation_id
  */
 BasicDealProduct.prototype['product_variation_id'] = undefined;
 /**
- * Any textual comment associated with this product-deal attachment. Visible and editable in the application UI.
+ * A textual comment associated with this product-deal attachment
  * @member {String} comments
  */
 BasicDealProduct.prototype['comments'] = undefined;
@@ -217,10 +234,11 @@ BasicDealProduct.prototype['comments'] = undefined;
  */
 BasicDealProduct.prototype['tax'] = 0;
 /**
- * Whether the product is enabled on the deal or not. This makes it possible to add products to a deal with a specific price and discount criteria - but keep them disabled, which refrains them from being included in the deal price calculation. When omitted, the product will be marked as enabled by default.
- * @member {module:model/NumberBoolean} enabled_flag
+ * Whether the product is enabled for a deal or not. This makes it possible to add products to a deal with a specific price and discount criteria, but keep them disabled, which refrains them from being included in the deal value calculation. When omitted, the product will be marked as enabled by default.
+ * @member {Boolean} enabled_flag
+ * @default true
  */
-BasicDealProduct.prototype['enabled_flag'] = undefined;
+BasicDealProduct.prototype['enabled_flag'] = true;